new-ui/requirements.txt
PyQt6>=6.4.0
python-dateutil>=2.8.0
pyswisseph>=2.10.0
geocoder>=1.38.1
geopy>=2.3.0
pytz>=2022.1
timezonefinder>=6.0.0
Pillow>=9.0.0
google-generativeai>=0.3.0
